ベンチマークのGitHub Actionを分離
